tests/wpt/web-platform-tests/html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-tokenization-noopener.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML: window.open `features`: tokenization -- `noopener`</title>
<meta name=timeout content=long>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#apis-for-creating-and-navigating-browsing-contexts-by-name">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
var windowURL = 'resources/close-self.html';

// Tests for how windows features are tokenized into 'name', 'value'
// window features separators are ASCII whitespace, '=' and  ','

test (t => {
  // Tokenizing `name`: initial window features separators are ignored
  // Each of these variants should tokenize to ('noopener', '')
  var featureVariants = [
    ' noopener',
    '=noopener',
    ',,noopener',
    ',=, noopener',
    '\n=noopener=',
    '\tnoopener',
    '\r,,,=noopener',
    '\u000Cnoopener'
  ];
  featureVariants.forEach(feature => {
    var win = window.open(windowURL, '', feature);
    assert_equals(win, null, `"${feature}" should activate feature "noopener"`);
  });
}, 'tokenization should skip window features separators before `name`');

test (t => {
  // Tokenizing `name`: lowercase conversion
  // Each of these variants should tokenize as feature ('noopener', '')
  // except where indicated
  // Note also that `value` is lowercased during tokenization
  var featureVariants = [
    'NOOPENER',
    'noOpenER',
    '  NOopener',
    '=NOOPENER',
    'noopener=1',
    'NOOPENER=1',
    'NOOPENER=yes',
    'noopener=YES',
  ];
  featureVariants.forEach(feature => {
    var win = window.open(windowURL, '', feature);
    assert_equals(win, null, `"${feature}" should activate feature "noopener"`);
  });
}, 'feature `name` should be converted to ASCII lowercase');

test (t => {
  // After `name` has been collected, ignore any window features separators until '='
  // except ',' OR a non-window-features-separator — break in those cases
  // i.e. ignore whitespace until '=' unless a ',' is encountered first
  // Each of these variants should tokenize as feature ('noopener', '')
  var featureVariants = [
    'noopener',
    ' noopener\r',
    'noopener\n =',
    'noopener,',
    'noopener  =,',
    ', noopener   =',
    'noopener,=',
    'noopener foo', // => ('noopener', ''), ('foo', '')
    'foo noopener=1', // => ('foo', ''), ('noopener', '1')
    'foo=\u000Cbar\u000Cnoopener' // => ('foo', 'bar'), ('noopener', '')
  ];
  featureVariants.forEach(feature => {
    var win = window.open(windowURL, '', feature);
    assert_equals(win, null, `"${feature}" should activate feature "noopener"`);
  });
}, 'after `name`, tokenization should skip window features separators that are not "=" or ","');

test (t => {
  // After initial '=', tokenizing should ignore all separators except ','
  // before collecting `value`
  // Each of these variants should tokenize as feature ('noopener', '')
  // Except where indicated
  var featureVariants = [
    'noopener=  yes', // => ('noopener', 'yes')
    'noopener==,',
    'noopener=\n ,',
    'noopener = \t ,',
    'noopener\n=\r 1,', // => ('noopener', '1')
    'noopener=,yes', // => ('noopener'), ('yes')
    'noopener= yes=,', // => ('noopener', 'yes')
    'noopener = \u000Cyes' // => ('noopener', 'yes')
  ];
  featureVariants.forEach(feature => {
    var win = window.open(windowURL, '', feature);
    assert_equals(win, null, `"${feature}" should activate feature "noopener"`);
  });
}, 'Tokenizing should ignore window feature separators except "," after initial "=" and before value');

test (t => {
  // Tokenizing `value` should collect any non-separator code points until first separator
  var featureVariants = [
    'noopener=1', // => ('noopener', 'noopener')
    'noopener=yes', // => ('noopener', 'yes')
    'noopener = yes ,', // => ('noopener', 'yes')
    'noopener=\nyes  ,', // => ('noopener', 'yes')
    'noopener=yes yes', // => ('noopener', 'yes'), ('yes', '')
    'noopener=yes\ts', // => ('noopener', 'yes'), ('s', '')
    'noopener==', // => ('noopener', '')
    'noopener=1\n,', // => ('noopener', '1')
    '==noopener===', // => ('noopener', '')
    'noopener==\u000C' // => ('noopener', '')
  ];
  featureVariants.forEach(feature => {
    var win = window.open(windowURL, '', feature);
    assert_equals(win, null, `"${feature}" should set "noopener"`);
  });
}, 'Tokenizing should read characters until first window feature separator as `value`');

test (t => {
  var featureVariants = [
    'noopener=1',
    'noopener=2',
    'noopener=12345',
    'noopener=1.5',
    'noopener=-1',
  ];
  featureVariants.forEach(feature => {
    var win = window.open(windowURL, '', feature);
    assert_equals(win, null, `"${feature}" should activate feature "noopener"`);
  });
}, 'Integer values other than 0 should activate the feature');

test (t => {
  var featureVariants = [
    'noopener=0',
    'noopener=0.5',
    'noopener=error',
  ];
  featureVariants.forEach(feature => {
    var win = window.open(windowURL, '', feature);
    assert_not_equals(win, null, `"${feature}" should NOT activate feature "noopener"`);
  });
}, 'Integer value of 0 should not activate the feature');

test (t => {
  var invalidFeatureVariants = [
    '-noopener', //     => ('-noopener', '')
    'NOOPENERRRR', //   => ('noopenerrr', '')
    'noOpenErR', //     => ('noopenerr', '')
    'no_opener', //     => ('no_opener', '')
    ' no opener', //    => ('no', ''), ('opener', '')
    'no\nopener', //    => ('no', ''), ('opener', '')
    'no,opener', //     => ('no', ''), ('opener', '')
    '\0noopener', //    => ('\0noopener', '')
    'noopener\u0000=yes', // => ('noopener\0', 'yes')
    'foo=\u000Cnoopener' // => ('foo', 'noopener')
  ];
  invalidFeatureVariants.forEach(feature => {
    var win = window.open(windowURL, '', feature);
    assert_not_equals(win, null, `"${feature}" should NOT activate feature "noopener"`);
  });
}, 'invalid feature names should not tokenize as "noopener"');
</script>
