LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-tokenization-noopener-expected.txt


PASS tokenization should skip window features separators before `name` 
FAIL feature `name` should be converted to ASCII lowercase assert_equals: "noopener=NOOPENER" should activate feature "noopener" expected null but got object "[object Window]"
PASS after `name`, tokenization should skip window features separators that are not "=" or "," 
FAIL Tokenizing should ignore window feature separators except "," after initial "=" and before value assert_equals: "noopener
=\r noopener," should activate feature "noopener" expected null but got object "[object Window]"
FAIL Tokenizing should read characters until first window feature separator as `value` assert_equals: "noopener=noopener" should set "noopener" expected null but got object "[object Window]"
FAIL "noopener" should be based on name (key), not value assert_equals: "noopener=false" should activate feature "noopener" expected null but got object "[object Window]"
PASS invalid feature names should not tokenize as "noopener" 
